% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_hex_list.R
\docType{data}
\name{cmap_hex_list}
\alias{cmap_hex_list}
\title{Hex codes}
\format{List of 43 different palettes containing 1000 hex codes
\describe{

}}
\source{
custom made
}
\usage{
cmap_hex_list
}
\description{
Hex codes
}
\examples{
if (require("EvoFreq")) {
cmap_hex_list
}
}
\keyword{datasets}
