% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.easy.long.edges.R
\docType{data}
\name{example.easy.long.edges}
\alias{example.easy.long.edges}
\title{Clone edges in Easy Long Format}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{parent}{Parent of clone}
  \item{clone}{Clone}
  \item{fitness}{Clone attribute that to color by}
}}
\source{
Custom Made.
}
\usage{
example.easy.long.edges
}
\description{
A dataset for easy understanding of EvoFreq data structures.
}
\examples{
?example.easy.long.edges
data("example.easy.long.edges")
}
\keyword{datasets}
