% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.easy.long.sizes.R
\docType{data}
\name{example.easy.long.sizes}
\alias{example.easy.long.sizes}
\title{Clone information dataset in Easy Long Format}
\format{A data frame with 47 rows and 3 variables:
\describe{
  \item{clone}{Clone}
  \item{Time}{Timepoint at which clone size is measured}
  \item{Size}{Size of clone at timepoint in Time}
}}
\source{
Custom Made.
}
\usage{
example.easy.long.sizes
}
\description{
A dataset for easy understanding of EvoFreq data structures.
}
\examples{
?example.easy.long.sizes
data("example.easy.long.sizes")
}
\keyword{datasets}
