% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.Easy.Wide.R
\docType{data}
\name{example.easy.wide}
\alias{example.easy.wide}
\title{Clone information dataset in Easy Wide Format}
\format{A data frame with 8 rows and 10 variables:
\describe{
  \item{parents}{Parent clone}
  \item{clones}{Clone metadata}
  \item{1}{Timepoint 1 clone sizes}
  \item{2}{Timepoint 2 clone sizes}
  \item{3}{Timepoint 3 clone sizes}
  \item{4}{Timepoint 4 clone sizes}
  \item{5}{Timepoint 5 clone sizes}
  \item{6}{Timepoint 6 clone sizes}
  \item{7}{Timepoint 7 clone sizes}
  \item{8}{Timepoint 8 clone sizes}
}}
\source{
Custom Made.
}
\usage{
example.easy.wide
}
\description{
A dataset for easy understanding of EvoFreq data structures.
}
\examples{
?example.easy.wide
data("example.easy.wide")
}
\keyword{datasets}
