% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.easy.wide.with.attributes.R
\docType{data}
\name{example.easy.wide.with.attributes}
\alias{example.easy.wide.with.attributes}
\title{Clone information dataset in Easy Wide Format}
\format{A data frame with 8 rows and 12 variables:
\describe{
  \item{parent}{Parent clone}
  \item{clone}{Clone metadata}
  \item{fitness}{Clone attribute to color by}
  \item{color}{Clone attribute that is custom color}
  \item{0}{Timepoint 0 clone sizes}
  \item{10}{Timepoint 10 clone sizes}
  \item{20}{Timepoint 20 clone sizes}
  \item{30}{Timepoint 30 clone sizes}
  \item{40}{Timepoint 40 clone sizes}
  \item{50}{Timepoint 50 clone sizes}
  \item{60}{Timepoint 60 clone sizes}
  \item{70}{Timepoint 70 clone sizes}
}}
\source{
Custom Made.
}
\usage{
example.easy.wide.with.attributes
}
\description{
A dataset for easy understanding of EvoFreq data structures.
}
\examples{
?example.easy.wide.with.attributes
data("example.easy.wide.with.attributes")
}
\keyword{datasets}
