% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.Long.Clone.History.R
\docType{data}
\name{example.long.clones}
\alias{example.long.clones}
\title{Clone information dataset in Long Format}
\format{A data frame with 2677 rows and 5 variables:
\describe{
  \item{clone_id}{ID of clone}
  \item{time}{Time of measurement}
  \item{size}{Size of clone}
}}
\source{
\url{https://www.biorxiv.org/content/10.1101/623959v2}
}
\usage{
example.long.clones
}
\description{
A dataset from a simulation taken from "Niche engineering drives early passage through an immune bottleneck in progression to colorectal cancer"
}
\examples{
if (require("EvoFreq")) {
example.long.clones
}
}
\keyword{datasets}
