% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.Long.Edges.R
\docType{data}
\name{example.long.edges}
\alias{example.long.edges}
\title{Clone information dataset in Long Format}
\format{A data frame with 1181 rows and 5 variables:
\describe{
  \item{parents}{Parent clone}
  \item{fitness}{Clone metadata}
  \item{drivers}{Driver mutations of clone}
  \item{passengers}{Passenger mutations of clone}
  \item{clone_id}{ID of the clone}
}}
\source{
\url{https://www.biorxiv.org/content/10.1101/623959v2}
}
\usage{
example.long.edges
}
\description{
A dataset from a simulation taken from "Niche engineering drives early passage through an immune bottleneck in progression to colorectal cancer"
}
\examples{
if (require("EvoFreq")) {
example.long.edges
}
}
\keyword{datasets}
