% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoFreq_funcs.R
\name{get_evofreq_labels}
\alias{get_evofreq_labels}
\title{get_evofreq_labels}
\usage{
get_evofreq_labels(freq_frame, apply_labels = FALSE,
  custom_label_text = NULL, evofreq_plot = NULL, clone_list = NULL,
  extant_only = FALSE, line_color = "darkgrey", adj.factor = 10)
}
\arguments{
\item{freq_frame}{Properly formatted dataframe from \code{\link{get_evofreq}}.}

\item{apply_labels}{Whether to plot the labels. Must include evofreq_plot if TRUE. Default FALSE.}

\item{custom_label_text}{Custom labels to use (e.g. c("Mutant X","Mutant Y"))}

\item{evofreq_plot}{Plot returned from \code{\link{plot_evofreq}}.}

\item{clone_list}{Only label clones with this id. Passed as a vector.}

\item{extant_only}{Get labels for only the extant clones.}

\item{line_color}{Single color value. Default "black".}

\item{adj.factor}{Value to scale the label and lines by so that they are far enough from the plot. Default value is 10 (a 10th of the distance between max/min at x from max/min across x).}
}
\value{
Dataframe containing the necessary coordinates to add labels in the proper locations.
}
\description{
get_evofreq_labels
}
\examples{
data("example.easy.wide") # Load Data
### Get frequency dataframe
freq_frame <- get_evofreq(example.easy.wide[,seq(3,10)], example.easy.wide$clones, example.easy.wide$parents, clone_cmap = "magma")
evofreq_plot <- plot_evofreq(freq_frame)
clone_labels <- get_evofreq_labels(freq_frame, extant_only=FALSE)

### Set evofreq_plot and apply_labels = TRUE to get a labeled plot back
get_evofreq_labels(freq_frame, extant_only=FALSE, evofreq_plot = evofreq_plot, apply_labels = TRUE)
}
