% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoFreq_Parsers.R
\name{parse_calder}
\alias{parse_calder}
\title{parse_calder

 Parse Calder outputs to visualize using EvoFreq.}
\usage{
parse_calder(soln, tree)
}
\arguments{
\item{soln}{Clonal composition matrix file (soln csv file).}

\item{tree}{dot file produced by Calder.}
}
\value{
A list of dataframes of resolved phylogenies from PhyloWGS.
}
\description{
parse_calder

 Parse Calder outputs to visualize using EvoFreq.
}
