% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoFreq_Parsers.R
\name{parse_phylowgs}
\alias{parse_phylowgs}
\title{parse_phylowgs

 Parse phylowgs outputs to visualize using EvoFreq.}
\usage{
parse_phylowgs(json_file, return_all = TRUE)
}
\arguments{
\item{json_file}{json file output from PhyloWGS containing tree structure and cellular prevalance information.}

\item{return_all}{if false, return the best result based on the linearity index provided by PhyloWGS.}
}
\value{
A list of dataframes of resolved phylogenies from PhyloWGS.
}
\description{
parse_phylowgs

 Parse phylowgs outputs to visualize using EvoFreq.
}
\examples{
\donttest{
phylowgs_output="run_name.summ.json"

tree_data <- parse_phylowgs(json_file=phylowgs_output)

#EvoFreqPlots
pdf('./evofreqs.pdf', width=8, height=4, onefile = T)
for (i in 1:length(f$all)){
 clone_dynamics_df <- get_evofreq(tree_data[[i]][,c(5,length(colnames(tree_data)))], clones=tree_data[[i]]$clone, parents=tree_data[[i]]$parent, clone_cmap = "jet")
 p <- plot_evofreq(evofreq_df)
 print(p)
}
dev.off()
}

}
