% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_funcs.R
\name{update_colors}
\alias{update_colors}
\title{update_colors

Update the colors of clones in the frequency dynamics dataframe or dendrogram plot dataframe}
\usage{
update_colors(evo_freq_df, attribute_df = NULL,
  attribute_val_name = NULL, clone_id_col_in_att_df = "clone_id",
  clone_cmap = "rainbow_soft", attribute_range = NULL)
}
\arguments{
\item{evo_freq_df}{Dataframe returned by \code{\link{get_evofreq}} or \code{\link{get_evogram}} , which contains the information to plot the frequency dynamics}

\item{attribute_df}{Optional dataframe containing attributes for each clone that can be used to determine the color of each clone in the plot. This can include a column of custom colors to use \code{attribute_val_name}.}

\item{attribute_val_name}{String defining which column in \code{attribute_df} will be used to color each clone. If attribute column values are a hex code in this column, these colors will be used instead of \code{clone_cmap}.}

\item{clone_id_col_in_att_df}{Name of the column in attribute_df that contains the clone IDs}

\item{clone_cmap}{String defining which colormap should be used to color the clones (nodes) if no attributes to color by. For a list of available colormaps, see \code{\link[colormap]{colormaps}}. If color not in  \code{\link[colormap]{colormaps}}, it is assumed all colors should be the same}

\item{attribute_range}{Range of values for the attribute to color by. If NULL, then range is determined from the attribute data}
}
\description{
update_colors

Update the colors of clones in the frequency dynamics dataframe or dendrogram plot dataframe
}
