% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoFreq_funcs.R
\name{plot_evofreq}
\alias{plot_evofreq}
\title{\code{plot_evofreq} Plots the frequency dynamics using ggplot. Can also use info to create animations}
\usage{
plot_evofreq(freq_frame, n_time_pts = NULL, start_time = NULL,
  end_time = NULL, bw = 0.05, bc = "grey75", show_axes = T)
}
\arguments{
\item{freq_frame}{Proper dataframe returned by \code{\link{get_evofreq}}}

\item{n_time_pts}{Integer defining how many time points to plot, evenly spaced between \code{start_time} and \code{end_time}. If NULL, all timepoints are plotted}

\item{start_time}{Integer defining the timepoint at which to start plotting frequency dynamics. If NULL, plotting will begin at the first timepoint}

\item{end_time}{Integer defining the timepoint at which to stop plotting frequency dynamics. If NULL, plotting will end at the last timepoint}

\item{bw}{width of lines surrounding polygons}

\item{bc}{color of lines surrounding polygons}

\item{show_axes}{Whether to display axes}
}
\value{
ggplot of the frequency dynamics
}
\description{
\code{plot_evofreq} Plots the frequency dynamics using ggplot. Can also use info to create animations
}
